requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
pep8==1.7.1
pycodestyle==2.10.0
Pygments==2.14.0
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.2
sqlparse==0.4.3
urllib3==1.26.15
psycopg2-binary==2.9.5
gunicorn==20.1.0
